+++ tests/test_requests.py
@@ -36,7 +36,7 @@ def test_request_with_json_to_log(log_builder):
     ],
 )
 def test_response_to_log(
-    log_builder: RequestsFullRequestLogBuilder, requests_request, response_params
+        log_builder: RequestsFullRequestLogBuilder, requests_request, response_params
 ):
     responses.add("POST", requests_request.url, **response_params)
 
@@ -62,7 +62,7 @@ def test_500_response_to_log(log_builder, requests_request):
 
 
 def test_full_request_to_log(
-    freezer, log_builder: RequestsFullRequestLogBuilder, requests_request, requests_response
+        freezer, log_builder: RequestsFullRequestLogBuilder, requests_request, requests_response
 ):
     actual_log = log_builder(requests_request, requests_response)
     expected_log = FullRequestLog(
@@ -72,15 +72,24 @@ def test_full_request_to_log(
     assert actual_log == expected_log
 
 
+def test_log_request_and_non_ok_response(
+        freezer, log_builder: RequestsFullRequestLogBuilder, requests_request, error_requests_response
+):
+    actual_log = log_builder(requests_request, error_requests_response)
+    expected_log = FullRequestLog(
+        log_builder.request_to_log(requests_request),
+        log_builder.response_to_log(error_requests_response),
+    )
+    assert actual_log == expected_log
+
+
 @pytest.mark.skipif(
     not config.ELASTIC_SEARCH_RUN_TESTS,
     reason="Set ELASTIC_SEARCH_RUN_TESTS env to enable Elastic Search tests",
 )
 def test_requests_log_insert_to_es(
-    freezer, test_es_index, requests_request, requests_response, serialized_full_request_log,
+        freezer, test_es_index, requests_request, requests_response, serialized_full_request_log,
 ):
     doc_id = log(requests_request, requests_response)
     actual_full_request_log = get_from_es(doc_id)
     assert actual_full_request_log == serialized_full_request_log
-
-